fix spec, reorganize file structure
