+++ packages/providers/src/index.ts
@@ -1,2 +1,3 @@
 export * from './types.js';
+export * from './error.js';
 export * from './ws/index.js';
