+++ packages/providers/src/smoldot/SmoldotProvider.ts
@@ -1,0 +1,11 @@
+import type { Chain } from 'smoldot';
+import { SubscriptionProvider } from '../base/index.js';
+
+/**
+ * @name SmoldotProvider
+ */
+export class SmoldotProvider extends SubscriptionProvider {
+  constructor(chain: Chain) {
+    super();
+  }
+}
